forecast_service/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
catboost==1.2.5
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
contourpy==1.2.1
coreforecast==0.0.12
cycler==0.12.1
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
graphviz==0.20.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lightning-utilities==0.11.6
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.3
neuralforecast==1.7.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
optuna==3.6.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.23.0
protobuf==5.27.2
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.3.3
pytz==2024.1
PyYAML==6.0.1
ray==2.32.0
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
scikit-learn==1.5.1
scipy==1.14.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
sympy==1.13.1
tenacity==8.5.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
torch==2.4.0
torchmetrics==1.4.0.post0
tqdm==4.66.4
triton==3.0.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
utilsforecast==0.2.0
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
